Script to check connectivity to azure
Azure use signature checking across provided headers. It would be
nice to have a tool to debug unexpected intermediate headers
modifications that may lead to errors like "The MAC signature found in
the HTTP request is not the same as any computed signature". Even
without access_key the script allow to show headers server use to
calculate the signature.

Signed-off-by: Petr Fedchenkov <giggsoff@gmail.com>