wwan: give CLI commands that got stuck a chance to exit cleanly
Behind the scenes, every QMI CLI command starts by allocating Client
ID (CID), then uses it for the duration of the request and releases it
just before returning.

However, we wrap qmicli calls with "timeout" to make sure that if
a command gets stuck (which happens quite often), it will not hang
indefinitely. But if a command timeouts and gets killed abruptly with
SIGKILL, it does not get the opportunity to release the allocated
client ID and CID will leak.

The number of available client IDs is quite limited - some services,
like NAS, may provide only 10. After we allocate (and leak) all of them,
we are no longer able to use the service (getting 'ClientIdsExhausted'
error) and a modem reboot is required to fix the issue.

Even when command is seemingly stuck, it will still actually react to
SIGINT and exit with all resources freed (I'm yet to experience a case
where SIGKILL was necessary). This commit therefore changes args for
the "timeout" command to send the SIGINT signal instead.
If the underlying command remains hanging still, SIGKILL will be sent
5 seconds later, avoiding an indefinite hang at the cost of one CID
leaked.

Signed-off-by: Milan Lenco <milan@zededa.com>