usbmanager: several fixes and small improvements
skip deleting VM:
if usb passthrough is removed, it does not mean that the
VM is gone, too.

handle error parsing uevent file:
The bufio.Scanner can fail, f.e. if the line is too long
for it's internal buffer.
In case this code runs in such an error, it is better to
just ignore the usb device instead of passing an
incomplete usb device and possibly pass it through to the
wrong edge application.

improve logging of pubsub:
log all pubsub events to ease debugging in case unexpected
behavior shows up.

Signed-off-by: Christoph Ostarek <christoph@zededa.com>