pkg/pillar: kvm does not call KvmContest:Delete
The handleDelete function in domainmgr does not call the Delete
function of the underlying hypervisor struct.  The result is that
the status directory (/run/hypervisor/kvm/<domainName>) is not
removed.

Adding the function then fails as the socket qmp has been deleted.
Added check in KvmContext:Delete to only try to remove this qemu
control socket if it exists.

Signed-off-by: Gerald (Bob) Lee <bob@famleehouse.net>