pkg/*/make-raw: check existence of IMX8 files in a separate function
! Patch covers two almost identical 'make-raw' files.

Check for IMX8 files repeats several times. Introduce a separate 'imx8_exists()'
function for this purpose.

Note! The 'make-raw' in the verification package is a bit outdated, so IMX8
blobs are copied to the /efifs folder. Originally this was introduced in
the following commit:

  b8f73a8af7ab ("arm: Add support to install EVE image on eMMC for i.MX8MP devices")

but was not ported to the verification 'make-raw'. This patch does that.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>