Add the EnableVncShimVm flag
We differentiate between the access to the edge application and the
shim VM:

- `enableVNC` is used to enable VNC for the edge application
- `enableVncShimVm` is used to enable VNC for the shim VM

This differentiation is need, because currently shim VM doesn't support
any user authentication, so anybody able to access the VNC port can
access the shim VM. This way VNC access to the shim VM stays disabled
by default unless explicitly enabled by the controller.

VNC is enabled in QEMU if either of the flags is set, but the
virtconsole is only enabled if the `enableVNCshimVm` flag is set.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>