Do not forget to report error severity
For network instances, network connectivity testing and IO bundle
processing we do not really use different error severity levels
at the moment.
For any encountered issue we simply report an error message up to
zedagent and further to the controller. However, zedagent forgets
to set the severity to "error" and leaves it as unspecified.
This may have unwanted effect on how the error message is
treated/displayed by the controller.

Later, we could improve this and use different severity levels also
for connectivity testing. For example, longer the port connectivity
is down, the more severe the error reported would be.

Signed-off-by: Milan Lenco <milan@zededa.com>