grub/rootfs.cfg: add 'isolate CPU0 (only for PREEMPT_RT)' menu option
This will set the 'isolcpus=inverse,0 nohz_full=inverse,0' kernel command
line, which by default pins all the service tasks running on EVE to the
CPU0, isolating other CPUs to VM needs and prevents scheduling interference.

This command line uses the 'inverse' flag, which was introduced here:

   https://github.com/lf-edge/eve-kernel/pull/90

Now the configuration is general and does not depend on the number of CPUs.
This option makes sense to set for the PREEMPT_RT kernel, which uses the
'inverse' flag.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>