000-kexec: remove 'nomodule' parameter, allows crashkernel to load modules
Patch removes 'nomodule' kernel parameter, which lets crashskernel
to load modules. This is utterly needed for interacting with the /persist
formatted as ZFS. ZFS is built separately out-of-tree as a module, so
has to be loaded explicitly by the crashkernel.

The majority of other modules are excluded from loading for the
crashkernel, see the 000-mod-params script.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>