runme.sh: check for content of /in
The user of EVE container can mount their own content into /in to change
/config of the resulting image. This change makes it possible to mount
single files instead of directories. If anything is present in /in - it
will be copied to /config.

`mountpoint -q /in` fails however if the mount path doesn't match "/in"
exactly. Another option would be to use `mount | grep "/in"`, but imo
`ls -A /in`` is simpler and cleaner, since we don't expect any other
files to be present there except the mounted ones.

Signed-off-by: Paul Gaiduk <paulg@zededa.com>