memory-monitor: Update README to reflect integration into rootfs.
Reflect the recent integration of the tool into the rootfs. The
documentation now includes details about how the tool is started as a
daemon on the EVE system, its location within a dedicated container, and
how its configuration and output files are managed. The process of
building and deploying the tool using a Dockerfile and Makefile is also
explained. The configuration section has been updated to reflect the new
location and naming of the configuration file

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>