memory-monitor: Handle newline character in string to number conversion.
Modifies the strtodec and strtoudec functions to handle the case where
the input string ends with a newline character. Previously, these
functions only considered a string valid if it ended with a null
character. Now, they also accept strings ending with a newline, which is
a common case when reading input from files or the console.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>